Move ajax select2 fields to dashboard code
