+++ saleor/dashboard/forms.py
@@ -1,6 +1,7 @@
 import json
 
 from django import forms
+from django.core.exceptions import ValidationError
 
 
 class AjaxSelect2ChoiceField(forms.ChoiceField):
